misc blog refactors
