--- src/components/Blog/index.ts
@@ -1,14 +1,0 @@
-// Blog Components
-export * from './BlogHeader'
-export * from './LinkPost'
-export * from './PodcastAudioPlayer'
-export * from './PodcastListItem'
-export * from './PostListItem'
-export * from './PostLayout'
-export * from './PostsList'
-export * from './PostsListPage'
-export * from './ReleasePost'
-export * from './BlogListItem'
-
-// Blog Post Data
-export * from './postTypes'
