+++ src/components/Blog/postTypes.tsx
@@ -7,13 +7,6 @@ export const BLOG_TYPE_TO_INFO: Record<BlogType, BlogTypeInfo> = {
     blog: {
         title: 'Blog',
         baseUrl: '/blog',
-        belowTitle: (
-            <>
-                <p className="mb-1 text-xl">
-                    Our changelog, announcements, dev posts, and anything else we think you'll find interesting.
-                </p>
-            </>
-        ),
         meta: {
             title: 'Sourcegraph blog',
             description: "Our changelog, announcements, dev posts, and anything else we think you'll find interesting.",
