+++ src/components/index.ts
@@ -31,21 +31,6 @@ export { YouTube } from './YouTube'
 export { StarshipPageCallToAction } from './cta/StarshipPageCallToAction'
 export { Banner } from './Banner'
 
-// Blog
-export {
-    BLOG_TYPE_TO_INFO,
-    BlogHeader,
-    LinkPost,
-    PodcastAudioPlayer,
-    PodcastListItem,
-    PostLayout,
-    PostListItem,
-    PostsList,
-    PostsListPage,
-    ReleasePost,
-    BlogListItem,
-} from './Blog'
-
 // Carousels
 export { CustomCarousel, TabCarousel, QuoteCarousel } from './Carousels'
